Address PR feedback for #848
